% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpu_temp.R
\name{get_cpu_temp_c}
\alias{get_cpu_temp_c}
\title{Get CPU temperature of Raspberry Pi in degree celsius}
\usage{
get_cpu_temp_c()
}
\value{

}
\description{
Get CPU temperature of Raspberry Pi in degree celsius
}
