% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storage.R
\name{get_storage_amount}
\alias{get_storage_amount}
\title{Returns dataframe with maximum storage, used storage, available storage
and % used storage}
\usage{
get_storage_amount()
}
\value{

}
\description{
Returns dataframe with maximum storage, used storage, available storage
and % used storage
}
