% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storage_pieplot.R
\name{plot_storage_pieplot}
\alias{plot_storage_pieplot}
\title{Create picharts for storage usage from rasppies}
\usage{
plot_storage_pieplot(data_set, is_hdd = TRUE)
}
\arguments{
\item{data_set}{Dataset provided by bash commands.}

\item{is_hdd}{Whether its HDD or RAM Storage}
}
\value{

}
\description{
Create picharts for storage usage from rasppies
}
